lazily load default logger if no logger provided
slogtripper was capturing the slog.Default logger on creation so if the default
logger changed between creation of slogtripper and the actual log being written,
slogtripper would use the original default not the new default logger. This made
it impossible to dynamically adjust log levels at runtime without recreating the
slogtripper instance (and potentially the whole http client/transport stack).